src/main/webapp/app/entities/assessment/service/assessment.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import dayjs from 'dayjs/esm';

import { DATE_TIME_FORMAT } from 'app/config/input.constants';
import { IAssessment, Assessment } from '../assessment.model';

import { AssessmentService } from './assessment.service';

describe('Assessment Service', () => {
  let service: AssessmentService;
  let httpMock: HttpTestingController;
  let elemDefault: IAssessment;
  let expectedResult: IAssessment | IAssessment[] | boolean | null;
  let currentDate: dayjs.Dayjs;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
    });
    expectedResult = null;
    service = TestBed.inject(AssessmentService);
    httpMock = TestBed.inject(HttpTestingController);
    currentDate = dayjs();

    elemDefault = {
      id: 0,
      assessmentDate: currentDate,
      lastModified: currentDate,
      lastModifiedBy: 'AAAAAAA',
    };
  });

  describe('Service methods', () => {
    it('should find an element', () => {
      const returnedFromService = Object.assign(
        {
          assessmentDate: currentDate.format(DATE_TIME_FORMAT),
          lastModified: currentDate.format(DATE_TIME_FORMAT),
        },
        elemDefault
      );

      service.find(123).subscribe(resp => (expectedResult = resp.body));

      const req = httpMock.expectOne({ method: 'GET' });
      req.flush(returnedFromService);
      expect(expectedResult).toMatchObject(elemDefault);
    });

    it('should create a Assessment', () => {
      const returnedFromService = Object.assign(
        {
          id: 0,
          assessmentDate: currentDate.format(DATE_TIME_FORMAT),
          lastModified: currentDate.format(DATE_TIME_FORMAT),
        },
        elemDefault
      );

      const expected = Object.assign(
        {
          assessmentDate: currentDate,
          lastModified: currentDate,
        },
        returnedFromService
      );

      service.create(new Assessment()).subscribe(resp => (expectedResult = resp.body));

      const req = httpMock.expectOne({ method: 'POST' });
      req.flush(returnedFromService);
      expect(expectedResult).toMatchObject(expected);
    });

    it('should update a Assessment', () => {
      const returnedFromService = Object.assign(
        {
          id: 1,
          assessmentDate: currentDate.format(DATE_TIME_FORMAT),
          lastModified: currentDate.format(DATE_TIME_FORMAT),
          lastModifiedBy: 'BBBBBB',
        },
        elemDefault
      );

      const expected = Object.assign(
        {
          assessmentDate: currentDate,
          lastModified: currentDate,
        },
        returnedFromService
      );

      service.update(expected).subscribe(resp => (expectedResult = resp.body));

      const req = httpMock.expectOne({ method: 'PUT' });
      req.flush(returnedFromService);
      expect(expectedResult).toMatchObject(expected);
    });

    it('should partial update a Assessment', () => {
      const patchObject = Object.assign(
        {
          lastModifiedBy: 'BBBBBB',
        },
        new Assessment()
      );

      const returnedFromService = Object.assign(patchObject, elemDefault);

      const expected = Object.assign(
        {
          assessmentDate: currentDate,
          lastModified: currentDate,
        },
        returnedFromService
      );

      service.partialUpdate(patchObject).subscribe(resp => (expectedResult = resp.body));

      const req = httpMock.expectOne({ method: 'PATCH' });
      req.flush(returnedFromService);
      expect(expectedResult).toMatchObject(expected);
    });

    it('should return a list of Assessment', () => {
      const returnedFromService = Object.assign(
        {
          id: 1,
          assessmentDate: currentDate.format(DATE_TIME_FORMAT),
          lastModified: currentDate.format(DATE_TIME_FORMAT),
          lastModifiedBy: 'BBBBBB',
        },
        elemDefault
      );

      const expected = Object.assign(
        {
          assessmentDate: currentDate,
          lastModified: currentDate,
        },
        returnedFromService
      );

      service.query().subscribe(resp => (expectedResult = resp.body));

      const req = httpMock.expectOne({ method: 'GET' });
      req.flush([returnedFromService]);
      httpMock.verify();
      expect(expectedResult).toContainEqual(expected);
    });

    it('should delete a Assessment', () => {
      service.delete(123).subscribe(resp => (expectedResult = resp.ok));

      const req = httpMock.expectOne({ method: 'DELETE' });
      req.flush({ status: 200 });
      expect(expectedResult);
    });

    describe('addAssessmentToCollectionIfMissing', () => {
      it('should add a Assessment to an empty array', () => {
        const assessment: IAssessment = { id: 123 };
        expectedResult = service.addAssessmentToCollectionIfMissing([], assessment);
        expect(expectedResult).toHaveLength(1);
        expect(expectedResult).toContain(assessment);
      });

      it('should not add a Assessment to an array that contains it', () => {
        const assessment: IAssessment = { id: 123 };
        const assessmentCollection: IAssessment[] = [
          {
            ...assessment,
          },
          { id: 456 },
        ];
        expectedResult = service.addAssessmentToCollectionIfMissing(assessmentCollection, assessment);
        expect(expectedResult).toHaveLength(2);
      });

      it("should add a Assessment to an array that doesn't contain it", () => {
        const assessment: IAssessment = { id: 123 };
        const assessmentCollection: IAssessment[] = [{ id: 456 }];
        expectedResult = service.addAssessmentToCollectionIfMissing(assessmentCollection, assessment);
        expect(expectedResult).toHaveLength(2);
        expect(expectedResult).toContain(assessment);
      });

      it('should add only unique Assessment to an array', () => {
        const assessmentArray: IAssessment[] = [{ id: 123 }, { id: 456 }, { id: 52122 }];
        const assessmentCollection: IAssessment[] = [{ id: 123 }];
        expectedResult = service.addAssessmentToCollectionIfMissing(assessmentCollection, ...assessmentArray);
        expect(expectedResult).toHaveLength(3);
      });

      it('should accept varargs', () => {
        const assessment: IAssessment = { id: 123 };
        const assessment2: IAssessment = { id: 456 };
        expectedResult = service.addAssessmentToCollectionIfMissing([], assessment, assessment2);
        expect(expectedResult).toHaveLength(2);
        expect(expectedResult).toContain(assessment);
        expect(expectedResult).toContain(assessment2);
      });

      it('should accept null and undefined values', () => {
        const assessment: IAssessment = { id: 123 };
        expectedResult = service.addAssessmentToCollectionIfMissing([], null, assessment, undefined);
        expect(expectedResult).toHaveLength(1);
        expect(expectedResult).toContain(assessment);
      });

      it('should return initial array if no Assessment is added', () => {
        const assessmentCollection: IAssessment[] = [{ id: 123 }];
        expectedResult = service.addAssessmentToCollectionIfMissing(assessmentCollection, undefined, null);
        expect(expectedResult).toEqual(assessmentCollection);
      });
    });
  });

  afterEach(() => {
    httpMock.verify();
  });
});
